Post deleted after remove non confirmed place
None verified users cannot login
